refactor: extract duplicate spot finder
