+++ test/duplicate_spot_detection_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:poker_analyzer/models/v2/training_pack_spot.dart';
 import 'package:poker_analyzer/models/v2/hand_data.dart';
 import 'package:poker_analyzer/models/v2/hero_position.dart';
-import 'package:poker_analyzer/screens/v2/training_pack_template_editor_screen.dart';
+import 'package:poker_analyzer/services/spot_duplicate_finder.dart';
 
 HandData _hand(String pos, String hero, String board) {
   final b = <String>[];
